+++ AYR/src/gl.h
@@ -10,6 +10,6 @@ namespace AYR
 		
 	};
 
-	AYR_API void drawLine(const Vector2f& p1, const Vector2f& p2, const Vector3f& col, img *image);
-	AYR_API void drawTriangle(const Vector2f& p1, const Vector2f& p2, const Vector2f& p3, const Vector3f& col, img* image);
+	AYR_API void drawLine(const Vector2i& p1, const Vector2i& p2, const Vector3f& col, img *image);
+	AYR_API void drawTriangle(const Vector2i& p1, const Vector2i& p2, const Vector2i& p3, const Vector3f& col, img* image);
 }
